requirements.txt
fastapi
sentence_transformers
faiss-cpu
python-dotenv
uvicorn
pandas
numpy
streamlit
aisuite
openai
groq
pymilvus
langchain
langchain-community
langchain_groq
langchain_openai